composition_player working smooth like butter. It was a mayhem
