+++ db/schema.rb
@@ -21,8 +21,8 @@ ActiveRecord::Schema.define(version: 2022_02_13_122753) do
     t.integer "rep_count"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.integer "memory_list_id"
     t.string "name"
+    t.integer "memory_list_id"
     t.index ["instrumental_id"], name: "index_compositions_on_instrumental_id"
     t.index ["user_id"], name: "index_compositions_on_user_id"
   end
